+++ animation_nodes/data_structures/gpencils/gp_stroke_data.py
@@ -1,6 +1,7 @@
 import textwrap
 from ... data_structures.color import Color
 from .. lists.base_lists import FloatList, Vector3DList, ColorList
+from ... data_structures.color import Color
 
 class GPStroke:
     def __init__(self, vertices = None, strengths = None, pressures = None,